requirements.txt
faker==22.6.0
pandas==2.2.0
numpy==1.26.4
google-cloud-bigquery==3.17.2
pyarrow==15.0.0
db-dtypes==1.2.0
streamlit==1.31.0
plotly==5.18.0
python-dotenv==1.0.1